feat: Add POSTComment & GETCommentList

+++ service/types/response/common.go
@@ -35,7 +35,7 @@ type Video struct {
 
 // 评论信息
 type Comment struct {
-	ID          int64  `json:"id"`          // 评论id
+	ID          uint   `json:"id"`          // 评论id
 	User        User   `json:"user"`        // 评论用户信息
 	Content     string `json:"content"`     // 评论内容
 	Create_Date string `json:"create_date"` // 评论发布日期，格式 mm-dd
